+++ packages/next/src/client/components/bailout-to-client-rendering.ts
@@ -1,5 +1,5 @@
 import { BailoutToCSRError } from '../../shared/lib/lazy-dynamic/bailout-to-csr'
-import { workAsyncStorage } from './work-async-storage.external'
+import { workAsyncStorage } from '../../server/app-render/work-async-storage.external'
 
 export function bailoutToClientRendering(reason: string): void | never {
   const workStore = workAsyncStorage.getStore()
