+++ packages/next/src/client/components/redirect.ts
@@ -1,4 +1,4 @@
-import { actionAsyncStorage } from './action-async-storage.external'
+import { actionAsyncStorage } from '../../server/app-render/action-async-storage.external'
 import { RedirectStatusCode } from './redirect-status-code'
 
 const REDIRECT_ERROR_CODE = 'NEXT_REDIRECT'
