+++ packages/next/src/server/app-render/clean-async-snapshot-instance.ts
@@ -1,4 +1,4 @@
-import { createSnapshot } from '../../client/components/async-local-storage'
+import { createSnapshot } from '../../server/app-render/async-local-storage'
 
 export const runInCleanSnapshot: <R, TArgs extends any[]>(
   fn: (...args: TArgs) => R,
