+++ packages/next/src/server/async-storage/with-work-store.ts
@@ -1,5 +1,5 @@
 import type { WithStore } from './with-store'
-import type { WorkStore } from '../../client/components/work-async-storage.external'
+import type { WorkStore } from '../../server/app-render/work-async-storage.external'
 import type { AsyncLocalStorage } from 'async_hooks'
 import type { IncrementalCache } from '../lib/incremental-cache'
 import type { RenderOpts } from '../app-render/types'
