+++ packages/next/src/server/node-environment-extensions/utils.tsx
@@ -1,4 +1,4 @@
-import { workAsyncStorage } from '../../client/components/work-async-storage.external'
+import { workAsyncStorage } from '../app-render/work-async-storage.external'
 import { workUnitAsyncStorage } from '../app-render/work-unit-async-storage.external'
 import { abortOnSynchronousDynamicDataAccess } from '../app-render/dynamic-rendering'
