+++ packages/next/src/server/web/adapter.ts
@@ -18,7 +18,7 @@ import {
   withWorkStore,
   type WorkStoreContext,
 } from '../async-storage/with-work-store'
-import { workAsyncStorage } from '../../client/components/work-async-storage.external'
+import { workAsyncStorage } from '../../server/app-render/work-async-storage.external'
 import { NEXT_ROUTER_PREFETCH_HEADER } from '../../client/components/app-router-headers'
 import { getTracer } from '../lib/trace/tracer'
 import type { TextMapGetter } from 'next/dist/compiled/@opentelemetry/api'
