+++ packages/next/src/server/web/spec-extension/adapters/request-cookies.ts
@@ -2,7 +2,7 @@ import type { RequestCookies } from '../cookies'
 
 import { ResponseCookies } from '../cookies'
 import { ReflectAdapter } from './reflect'
-import { workAsyncStorage } from '../../../../client/components/work-async-storage.external'
+import { workAsyncStorage } from '../../../../server/app-render/work-async-storage.external'
 
 /**
  * @internal
