+++ packages/next/src/server/web/spec-extension/revalidate.ts
@@ -4,7 +4,7 @@ import {
   NEXT_CACHE_IMPLICIT_TAG_ID,
   NEXT_CACHE_SOFT_TAG_MAX_LENGTH,
 } from '../../../lib/constants'
-import { workAsyncStorage } from '../../../client/components/work-async-storage.external'
+import { workAsyncStorage } from '../../../server/app-render/work-async-storage.external'
 import { workUnitAsyncStorage } from '../../../server/app-render/work-unit-async-storage.external'
 
 /**
