+++ packages/next/src/server/web/spec-extension/unstable-cache.ts
@@ -6,7 +6,7 @@ import {
   validateRevalidate,
   validateTags,
 } from '../../lib/patch-fetch'
-import { workAsyncStorage } from '../../../client/components/work-async-storage.external'
+import { workAsyncStorage } from '../../../server/app-render/work-async-storage.external'
 import { workUnitAsyncStorage } from '../../app-render/work-unit-async-storage.external'
 import {
   CachedRouteKind,
